Move active storage related code to attachable concern
